FrontEnd/requirements.txt
SQLAlchemy==1.4.40
Flask
gunicorn==20.1.0
pandas
requests
werkzeug.security
